Move postcode ingestion tests to separate file
